updated attack.txt with xss and csrf attack ideas
XSS: stored or reflected attack.

Easiest to demonstrate is a stored xss attack where the attacker stores
the xss script on the targeted server through some vulnerability.

Our example could be the userlist with the xss being the username itself.

A reflected attack is trickier but could perhaps be managed by modifiying
the shopping_cart.php page to return some vulnerable page when you try to
add a product that doesn't exist.

CSRF:

Demonstrate by including an image with a GET request to shopping_cart.php
and later with a POST or GET request to checkout.php

Idea is that we want to change some state and not really try to take over
a user's session (as what we usually want in a XSS).