add csrf attack example
To see it in effect:

 * open two browser windows (same browser, though)

 * browser window 1: log in and add some products to your shopping cart

 * browser window 1: view your shopping cart

 * browser window 2: open the csrf_attack.html

 * browser window 1: refresh and be surprised

What actually happened was that in csrf_attack.html we avoid the
cross-domain rule and make a request to /receipt.php (which essentially is
our checkout page).

Because cookies are stored per-domain (with some path restrictions), the
session cookie is transmitted, even though we are on another 'website'.