add csrf.php script to protect receipt.php
the csrf.php script exposes 2 functions:

 * init_csrf_token(form_name) --> token

 * validate_csrf(form_name, recv_token) --> true/false

the idea is that we call `init_csrf_token` when rendering an input form,
and validate the request later on using `validate_csrf`.

init_csrf_token(form_name) takes a unique form name and generates a random
csrf token, stores that token in `$_SESSION[$form_name]` and returns it.

The typical use case is to embed the returned token in a hidden input
field "csrf_token", i.e:

	<input type="hidden" name="csrf_token" value=".......">

Another use case would be to manipulate the url of a link:

	<a href="receipt.php?csrf_token=.......">checkout</a>

When server parses the request it can then validate using:

	$token = $_POST["csrf_token"]; // or $_GET["csrf_token"];
	$valid = validate_csrf("the_form", $token);

and detect if the request is invalid.

The receipt.php uses the POST + <form> approach.

receipt_unsafe.php is a copy of the old receipt.php