add reflected xss html and make shop.php insecure
the reflected_xss.html is based around the idea that I give you a link
with some embedded javascript, you click it, and the targeted address (href)
returns something that will execute my embedded javascript.

In this case shop.php is set up to allow users to add products to their
shopping carts via GET, like this:

	http://localhost/websec/shop.php?add=some-product-here

If the product specified by add doesn't exist, we output an error msg.

To see it in effect:
 * open a window and log in to the websec shop.

 * open the reflected_xss.html in another tab/window.

 * watch how your sessionid is displayed :D

The vulnerability is in the way we treat/use the input data.
If we write:

	echo 'error: '.$add.' is not valid!';

the provided data will be interpreted as html, which allows for all kinds
of exploits. To fix this we usually sanitize user input:

	echo 'error: '.htmlspecialchars($add).' is not valid!';

(or something equivalent).